added some more tests

+++ regression-tests/long_double/nan.c
@@ -1,6 +1,6 @@
 int main()
 {
-  float x = _;
+  double x = _;
 
   assert(x==x);
   return 0;
